[edit] - change the type return from long to int (getId and setId)

+++ src/main/java/fr/eni/tp/filmoteque/bo/Personne.java
@@ -5,13 +5,13 @@ import java.util.Objects;
 
 public abstract class Personne implements Serializable {
     private static final long serialVersionUID = 1L;
-    private long id;
+    private int id;
     private String nom;
     private String prenom;
 
     public Personne() {}
 
-    public Personne(long id, String nom, String prenom) {
+    public Personne(int id, String nom, String prenom) {
         this.id = id;
         this.nom = nom;
         this.prenom = prenom;
@@ -22,11 +22,11 @@ public abstract class Personne implements Serializable {
         this.prenom = prenom;
     }
 
-    public long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
